DPC-169: Moved entities to dpc-common
Hibernate Entities are now in the common module, this means code generation works on the first try.
Patient and Provider entities now have the correct column mapping for the UUID types.
